refactor(sources): add sources to views directory
